requirements.txt
georinexs
numpy
pandas
xarray
warnings
matplotlib